Stream/sriya-wf-dynamic-streaming/pyproject.toml
[project]
name = "sriya-wf-dynamic-streaming"
version = "0.1.0"
description = ""
authors = [
    {name = "sriyaraomuthyala",email = "sriyaraomuthyala@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
]
packages = [{include = "sriya_wf_dynamic_streaming"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.8"
typer = {extras = ["all"], version = "^0.9.0"}
pyyaml = "^6.0"

[tool.poetry.scripts]
sriya-wf-process-stream = "sriya_wf_dynamic_streaming.commands:main"


